mavlink: move OBSTACLE_DISTANCE stream to separate file
